+++ packages/image_picker/image_picker_android/android/src/main/java/io/flutter/plugins/imagepicker/ImagePickerDelegate.java
@@ -35,6 +35,7 @@ import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.stream.Collectors;
 
 /**
  * A delegate class doing the heavy lifting for the plugin.
@@ -81,6 +82,7 @@ public class ImagePickerDelegate
   @VisibleForTesting static final int REQUEST_CAMERA_IMAGE_PERMISSION = 2345;
   @VisibleForTesting static final int REQUEST_CODE_CHOOSE_MULTI_IMAGE_FROM_GALLERY = 2346;
   @VisibleForTesting static final int REQUEST_CODE_CHOOSE_MEDIA_FROM_GALLERY = 2347;
+  @VisibleForTesting static final int REQUEST_CODE_CHOOSE_MULTI_VIDEO_FROM_GALLERY = 2348;
 
   @VisibleForTesting static final int REQUEST_CODE_CHOOSE_VIDEO_FROM_GALLERY = 2352;
   @VisibleForTesting static final int REQUEST_CODE_TAKE_VIDEO_WITH_CAMERA = 2353;
@@ -474,6 +476,38 @@ public class ImagePickerDelegate
         pickMultiImageIntent, REQUEST_CODE_CHOOSE_MULTI_IMAGE_FROM_GALLERY);
   }
 
+  public void chooseMultiVideoFromGallery(
+      @NonNull VideoSelectionOptions options,
+      boolean usePhotoPicker,
+      int limit,
+      @NonNull Messages.Result<List<String>> result) {
+    if (!setPendingOptionsAndResult(null, options, result)) {
+      finishWithAlreadyActiveError(result);
+      return;
+    }
+
+    launchMultiPickVideoFromGalleryIntent(usePhotoPicker, limit);
+  }
+
+  private void launchMultiPickVideoFromGalleryIntent(Boolean usePhotoPicker, int limit) {
+    Intent pickMultiVideoIntent;
+    if (usePhotoPicker) {
+      pickMultiVideoIntent =
+          new ActivityResultContracts.PickMultipleVisualMedia(limit)
+              .createIntent(
+                  activity,
+                  new PickVisualMediaRequest.Builder()
+                      .setMediaType(ActivityResultContracts.PickVisualMedia.VideoOnly.INSTANCE)
+                      .build());
+    } else {
+      pickMultiVideoIntent = new Intent(Intent.ACTION_GET_CONTENT);
+      pickMultiVideoIntent.setType("video/*");
+      pickMultiVideoIntent.putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true);
+    }
+    activity.startActivityForResult(
+        pickMultiVideoIntent, REQUEST_CODE_CHOOSE_MULTI_VIDEO_FROM_GALLERY);
+  }
+
   public void takeImageWithCamera(
       @NonNull ImageSelectionOptions options, @NonNull Messages.Result<List<String>> result) {
     if (!setPendingOptionsAndResult(options, null, result)) {
@@ -617,6 +651,9 @@ public class ImagePickerDelegate
       case REQUEST_CODE_CHOOSE_MULTI_IMAGE_FROM_GALLERY:
         handlerRunnable = () -> handleChooseMultiImageResult(resultCode, data);
         break;
+      case REQUEST_CODE_CHOOSE_MULTI_VIDEO_FROM_GALLERY:
+        handlerRunnable = () -> handleChooseMultiVideoResult(resultCode, data);
+        break;
       case REQUEST_CODE_TAKE_IMAGE_WITH_CAMERA:
         handlerRunnable = () -> handleCaptureImageResult(resultCode);
         break;
@@ -689,6 +726,25 @@ public class ImagePickerDelegate
       }
 
       handleMediaResult(paths);
+      return;
+    }
+
+    // User cancelled choosing a picture.
+    finishWithSuccess(null);
+  }
+
+  private void handleChooseMultiVideoResult(int resultCode, Intent intent) {
+    if (resultCode == Activity.RESULT_OK && intent != null) {
+      ArrayList<MediaPath> paths = getPathsFromIntent(intent, false);
+      // If there's no valid Uri, return an error
+      if (paths == null) {
+        finishWithError(
+            "missing_valid_video_uri", "Cannot find at least one of the selected videos.");
+        return;
+      }
+
+      finishWithListSuccess(
+          paths.stream().map(p -> p.getPath()).collect(Collectors.toCollection(ArrayList::new)));
       return;
     }