+++ packages/image_picker/image_picker_platform_interface/lib/src/platform_interface/image_picker_platform.dart
@@ -322,6 +322,21 @@ abstract class ImagePickerPlatform extends PlatformInterface {
     return pickedImages ?? <XFile>[];
   }
 
+  /// Returns a [List<XFile>] with the videos that were picked.
+  ///
+  /// The `source` argument controls where the video comes from. This can
+  /// be either [ImageSource.camera] or [ImageSource.gallery].
+  ///
+  /// The `options` argument controls additional settings that can be used when
+  /// picking a video. See [VideoPickerOptions] for more details.
+  ///
+  /// If no videos were picked, returns an empty list.
+  Future<List<XFile>> getMultiVideoWithOptions({
+    MultiVideoPickerOptions options = const MultiVideoPickerOptions(),
+  }) {
+    throw UnimplementedError('getMultiVideoWithOptions() has not been implemented.');
+  }
+
   /// Returns true if the implementation supports [source].
   ///
   /// Defaults to true for the original image sources, `gallery` and `camera`,