+++ packages/image_picker/image_picker/CHANGELOG.md
@@ -1,5 +1,10 @@
-## NEXT
+## 1.3.0
 
+* feat: Add pickMultiVideo.
+
+## 1.2.0
+
+* feat: Add pickMultiVideo.
 * Updates minimum supported SDK version to Flutter 3.27/Dart 3.6.
 
 ## 1.1.2